unit test structure tidy up, added WP8 test project and folders and naming updates
