+++ MarkerMetro.Unity.WinLegacyMetro.Tests/MarkerMetro.Unity.WinLegacyMetro.Tests.csproj
@@ -113,14 +113,19 @@
     <SDKReference Include="TestPlatform, Version=11.0" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="..\MarkerMetro.Unity.WinLegacyUnity.Tests\ReflectionTests.cs">
-      <Link>ReflectionTests.cs</Link>
+    <Compile Include="..\MarkerMetro.Unity.WinLegacyUnity.Tests\System\Collections\SortedListTests.cs">
+      <Link>System\Collections\SortedListTests.cs</Link>
     </Compile>
-    <Compile Include="..\MarkerMetro.Unity.WinLegacyUnity.Tests\TcpClientTests.cs">
-      <Link>TcpClientTests.cs</Link>
+    <Compile Include="..\MarkerMetro.Unity.WinLegacyUnity.Tests\System\IO\IOTests.cs">
+      <Link>System\IO\IOTests.cs</Link>
+    </Compile>
+    <Compile Include="..\MarkerMetro.Unity.WinLegacyUnity.Tests\System\Net\Sockets\TcpClientTests.cs">
+      <Link>System\Net\Sockets\TcpClientTests.cs</Link>
+    </Compile>
+    <Compile Include="..\MarkerMetro.Unity.WinLegacyUnity.Tests\System\Reflection\ReflectionTests.cs">
+      <Link>System\Reflection\ReflectionTests.cs</Link>
     </Compile>
     <Compile Include="NameValueCollectionTests.cs" />
-    <Compile Include="IOTests.cs" />
     <Compile Include="StreamReaderExtensionsTests.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
