+++ MarkerMetro.Unity.WinLegacyMetro/MarkerMetro.Unity.WinLegacyMetro.csproj
@@ -211,8 +211,8 @@
     <Compile Include="..\MarkerMetro.Unity.WinLegacyUnity\System\Net\WebRequest.cs">
       <Link>System\Net\WebRequest.cs</Link>
     </Compile>
-    <Compile Include="..\MarkerMetro.Unity.WinLegacyUnity\System\ReflectionExtensions.cs">
-      <Link>System\ReflectionExtensions.cs</Link>
+    <Compile Include="..\MarkerMetro.Unity.WinLegacyUnity\System\Reflection\ReflectionExtensions.cs">
+      <Link>System\Reflection\ReflectionExtensions.cs</Link>
     </Compile>
     <Compile Include="..\MarkerMetro.Unity.WinLegacyUnity\System\Reflection\TargetException.cs">
       <Link>TargetException.cs</Link>
@@ -236,9 +236,7 @@
     <Compile Include="System\IO\EncryptedStreamReader.cs" />
     <Compile Include="System\IO\EncryptedStreamWriter.cs" />
   </ItemGroup>
-  <ItemGroup>
-    <Folder Include="System\Reflection\" />
-  </ItemGroup>
+  <ItemGroup />
   <PropertyGroup Condition=" '$(VisualStudioVersion)' == '' or '$(VisualStudioVersion)' &lt; '12.0' ">
     <VisualStudioVersion>12.0</VisualStudioVersion>
   </PropertyGroup>
