+++ MarkerMetro.Unity.WinLegacyUnity.Tests/MarkerMetro.Unity.WinLegacyUnity.Tests.csproj
@@ -50,10 +50,11 @@
     </Otherwise>
   </Choose>
   <ItemGroup>
-    <Compile Include="TcpClientTests.cs" />
+    <Compile Include="System\IO\IOTests.cs" />
+    <Compile Include="System\Net\Sockets\TcpClientTests.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="ReflectionTests.cs" />
-    <Compile Include="SortedListTests.cs" />
+    <Compile Include="System\Reflection\ReflectionTests.cs" />
+    <Compile Include="System\Collections\SortedListTests.cs" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\MarkerMetro.Unity.WinLegacyUnity\MarkerMetro.Unity.WinLegacyUnity.csproj">
